+++ Additional/Teams/Team 4/src/cpp/Motorcycle.cpp
@@ -8,7 +8,7 @@ Motorcycle::Motorcycle() : GroundVehicle()
 	frameType = nullptr;
 }
 
-Motorcycle::Motorcycle(int maxPassengers, int maxSpeed, const String& name, int year, const String& color, size_t doorsCount,
+Motorcycle::Motorcycle(size_t maxPassengers, size_t maxSpeed, const String& name, size_t year, const String& color, size_t doorsCount,
 	const String& brand, const String& model, const String& frameType) : GroundVehicle(maxPassengers, maxSpeed, name, year, color, doorsCount)
 {
 	t = GroundType::MOTORCYCLE;
