+++ Additional/Teams/Team 4/src/cpp/Truck.cpp
@@ -5,7 +5,7 @@ Truck::Truck() : GroundVehicle()
 
 }
 
-Truck::Truck(int maxPassengers, int maxSpeed, const String& name, int year, const String& color, size_t doorsCount,
+Truck::Truck(size_t maxPassengers, size_t maxSpeed, const String& name, size_t year, const String& color, size_t doorsCount,
 	const String& brand, size_t capacity) : GroundVehicle(maxPassengers, maxSpeed, name, year, color, doorsCount)
 {
 	this->brand = brand;
