+++ Additional/Teams/Team 4/src/h/Car.h
@@ -9,7 +9,7 @@ private:
 	String fuelType;
 public:
 	Car();
-	Car(int maxPassengers, int maxSpeed, const String& name, int year, const String& color, size_t doorsCount,
+	Car(size_t maxPassengers, size_t maxSpeed, const String& name, size_t year, const String& color, size_t doorsCount,
 		const String& brand, const String& model, const String& fuelType);
 
 	Vehicle* clone() const override;
