+++ Additional/Teams/Team 4/src/h/GroundVehicle.h
@@ -17,7 +17,7 @@ private:
 	size_t doorsCount;
 public:
 	GroundVehicle();
-	GroundVehicle(int, int, const String&, int, const String&, size_t);
+	GroundVehicle(size_t, size_t, const String&, size_t, const String&, size_t);
 
 	void setColor(const String&);
 	const String& getColor() const;
